requirements.txt
altgraph==0.17
et-xmlfile==1.0.1
future==0.18.2
jdcal==1.4.1
openpyxl==3.0.5
pefile==2019.4.18
pyodbc==4.0.30
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
six==1.15.0
